Prevent redirects to malformed URLs
Any "proceed_to" value must be a well-formed URL or an absolute
path. Anything else will raise a RedirectProtector::Violation.

Previously, a malformed URL would pass the "ensure_same_origin" test,
providing a potential vector for a malicious URL (which obviously must
be chained with a cookie attack, since that's where the flash values
is persisted).